requirements.txt
fastapi 
strawberry-graphql[debug-server]
strawberry-graphql[fastapi]
uvicorn
sqlmodel
